requirements.txt
langchain==0.3.27
langchain-chroma==0.2.6
langchain-community==0.3.29
langchain-core==0.3.76
langchain-deepseek==0.1.4
langchain-huggingface==0.3.1
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
chromadb==1.1.0
unstructured==0.18.14
markdown==3.9
python-dotenv==1.1.1
streamlit==1.50.0
tiktoken==0.11.0
torch==2.7.1
transformers==4.56.1
openai==1.107.3